+++ nrf_fuel_gauge/CHANGELOG.rst
@@ -9,6 +9,30 @@ Changelog
 
 All notable changes to this project are documented on this page.
 
+Main branch
+***********
+
+Changes
+=======
+
+* Updated to library version 1.0.0.
+* Simplified the :c:func:`nrf_fuel_gauge_process` function, as vbus state information is now gathered through the :c:func:`nrf_fuel_gauge_ext_state_update` function.
+* Simplified the :c:func:`nrf_fuel_gauge_ttf_get` function, as charger state information is now gathered through the :c:func:`nrf_fuel_gauge_ext_state_update` function.
+* Simplified the use of the :c:func:`nrf_fuel_gauge_param_adjust` function by allowing parameters to be set to "nan" to indicate that they should not be adjusted.
+
+Added
+=====
+
+* Support for primary cell (non-recheargable) batteries.
+* Battery models for Alkaline AA, Alkaline AAA, Alkaline coin cell LR44, and Lithium-manganese dioxide coin cell CR2032 batteries.
+* A new :c:func:`nrf_fuel_gauge_ext_state_update` function to inform the library about external state changes.
+  Such changes includes charger state, vbus connection status, and current measurement correction.
+* A functionality to retrieve and restore the library state, using the :c:func:`nrf_fuel_gauge_state_get` function and state member of the :c:struct:`nrf_fuel_gauge_init_parameters` structure.
+* A new member in the :c:struct:`nrf_fuel_gauge_state_info` structure to determine the current measurement value used after corrections.
+* A new member in the :c:struct:`nrf_fuel_gauge_state_info` structure to determine the state-of-charge value before it is rounded off.
+  The rounding off only applies to primary cell batteries.
+* New undocumented members in the :c:struct:`nrf_fuel_gauge_config_parameters` structure, intended for future adjustability.
+
 nRF Connect SDK v2.8.0
 **********************